requirements.txt
numpy 
opencv-python 
Pillow 
selenium 
selenium-wire 
python-socketio 
requests
